添加测试交互题代码的模板 Add test templates for interactive problem

+++ copypasta/template/interactive_problem/main.go
@@ -3,7 +3,7 @@ package main
 import . "fmt"
 
 // github.com/EndlessCheng/codeforces-go
-func run(io func(int) bool) (ans int) {
+func run(io func(int64) bool) (ans int64) {
 	var n int
 	Scan(&n)
 
@@ -11,7 +11,7 @@ func run(io func(int) bool) (ans int) {
 }
 
 func main() {
-	io := func(q int) bool {
+	io := func(q int64) bool {
 		Println("?", q)
 		// ... or read int and return it
 		var s []byte
